chore: edit request
push users to form when editing, refactor tabs to page-partials folder
