+++ app/dispatchers/requestDispatcher.ts
@@ -6,16 +6,16 @@ export const $setRequests = (payload: Request[]): Action => {
   return { type: 'setRequests', payload };
 };
 
-export const $setEditingRequest = (payload: boolean): Action => {
-  return { type: 'setEditingRequest', payload };
-};
-
 export const $updateRequest = (payload: Request | null): Action => {
   return { type: 'updateRequest', payload };
 };
 
 export const $deleteRequest = (payload: number | null): Action => {
   return { type: 'deleteRequest', payload };
+};
+
+export const $setRequestToDelete = (payload: number | null): Action => {
+  return { type: 'setRequestIdToDelete', payload };
 };
 
 export const $setTeams = (payload: Team[] | null): Action => {
